+++ sopa/segmentation/methods.py
@@ -84,12 +84,13 @@ def comseg_patch(temp_dir: str, patch_index: int, config: dict):
     import json
 
     try:
+        import comseg
         from comseg import dataset as ds
         from comseg import dictionary
     except ModuleNotFoundError:
-        raise ModuleNotFoundError(
-            "Install the ComSeg package (`pip install comseg`) for this method to work"
-        )
+        raise ModuleNotFoundError("Install comseg (`pip install comseg`) for this method to work")
+
+    assert comseg.__version__ >= "1.3", "comseg version should be >= 1.3"
 
     path_dataset_folder = Path(temp_dir) / str(patch_index)
 
@@ -101,6 +102,8 @@ def comseg_patch(temp_dir: str, patch_index: int, config: dict):
         image_csv_files=["transcripts.csv"],
         centroid_csv_files=["centroids.csv"],
         path_cell_centroid=path_dataset_folder,
+        min_nb_rna_patch=config.get("min_nb_rna_patch", 0),
+        prior_name=config.get("prior_name", SopaKeys.DEFAULT_CELL_KEY),
     )
 
     dataset.compute_edge_weight(config=config)
@@ -108,7 +111,6 @@ def comseg_patch(temp_dir: str, patch_index: int, config: dict):
     Comsegdict = dictionary.ComSegDict(
         dataset=dataset,
         mean_cell_diameter=config["mean_cell_diameter"],
-        prior_name=SopaKeys.DEFAULT_CELL_KEY,
     )
 
     Comsegdict.run_all(config=config)